refactor: 抽取光标位置计算函数
